#278: Split rendering of notes into own component
